+++ CHANGELOG.md
@@ -9,6 +9,9 @@ new, required `fname` `portlet-preference` in Benefit Information Portlet.
 
 Putatively 9.2.1, thefore currently building 9.2.1-SNAPSHOT.
 
++ Breaking change:
+  Removes undocumented,
+  not-working-as-expected "tab" parameter feature from Benefit Information.
 + In Benefit Information,
   now supports portlet request parameter `requestedContent`,
   as in Payroll Information. Values "ETF WRS Statements of Benefits" and